backend/requirements.txt
fastapi
uvicorn[standard]
pandas
httpx
numpy==1.26.4
pillow

python-multipart
pydantic
pydantic-settings

requests
duckduckgo-search

python-dotenv

azure-cognitiveservices-vision-customvision

